Refactored backend dockerfile and __init__ file, added docker compose file (WIP)
