+++ Dockerfile
@@ -1,31 +1,10 @@
-FROM node:20 AS build-stage
+FROM python:3.11.7-alpine3.19
 
-WORKDIR /react-app
-COPY react-app/. .
+RUN --mount=type=bind,source=requirements.txt,target=requirements.txt \
+    pip install -r requirements.txt
 
-# Set REACT_APP_BASE_URL during build time
-# ENV REACT_APP_BASR_URL=https://your-app-name.deployment-site.com/
+COPY /flask_app .
 
-# Build React app
-RUN npm install
-RUN npm run build
+EXPOSE 5000
 
-FROM python:3.11.3
-
-# Set up Flask environment
-ENV FLASK_APP=flask_app
-ENV FLASK_ENV=production
-ENV SQLALCHEMY_ECHO=True
-
-EXPOSE 8000
-
-WORKDIR /var/www
-COPY . .
-COPY --from=build-stage /react-app/* app/static
-
-# Install Python dependencies
-RUN pip install -r requirements.txt
-RUN pip install psycopg2
-
-# Run Flask environment
-CMD gunicorn app:app
+CMD ["flask", "run"]
