+++ docker-compose.yml
@@ -1,0 +1,5 @@
+services:
+    backend:
+      build: .
+      ports:
+          - "5000:5000"
